+++ front/groupomania/src/containers/UserDelete.js
@@ -5,7 +5,6 @@ import { Fragment } from "react";
 import { useHistory } from "react-router-dom";
 
 function UserDelete({ userId, deleteUser }) {
-
   let history = useHistory();
 
   const userData = JSON.parse(localStorage.getItem("userData"));
@@ -16,7 +15,13 @@ function UserDelete({ userId, deleteUser }) {
         className="btn btn-outline-danger btn-sm"
         onClick={(e) => {
           e.preventDefault();
-          deleteUser(userId, history);
+          if (
+            window.confirm(
+              "Êtes-vous sûr de vouloir supprimer votre compte ? Celà entrainera également la suppression de tous vos posts, commentaires et likes."
+            )
+          ) {
+            deleteUser(userId, history);
+          }
         }}
       >
         <i className="far fa-trash-alt"></i>
